Sorts strongly associated scripts into subdirs and removes v old unused scripts
This should make it easier to make the full switch later, and the module is being changed accordingly to make the PATHs still work.

I've done some checks, but it's *possible* something could break if it's using an absolute path to one of the relocated scripts.

Note that by "strongly associated", I mean "this is closely connected to a particular subsystem and would need to be wholly reimplemented for a new system if it's even still relevant".

The two removed scripts were for copying data on previous filesystem transitions, and are no longer relevant.